requirements-dev.txt
-r requirements.txt
black==22.10.0
isort==5.10.1
mypy==0.982
pytest==7.1.3
envdir==1.0.1
pytest-django==4.5.2
faker==15.3.2
pytest-factoryboy==2.5.0